Focused on NPC/Dialogs and Dungeon Generator - Implemented basic dialog system - Started on NPC class and created interactables - Started Bartender NPC

+++ Scripts/Level/EnemySpawner.cs
@@ -7,8 +7,8 @@ public partial class EnemySpawner : Node
 {
 	[Export] public EnemyResource Enemy;
 	[Export] public float SpawnChance;
-	[Export] public float MinAmount;
-	[Export] public float MaxAmount;
+	[Export] public int MinAmount;
+	[Export] public int MaxAmount;
 
 	// Called when the node enters the scene tree for the first time.
 	public override void _Ready()
